+++ components/mailbox/CHANGELOG.md
@@ -27,6 +27,7 @@
 - Pass z-index css to confirmation pop up's modal & overlay using css variables `z-index-modal` & `z-index-overlay` respectively
 - Draft, reply, reply all and forwarded messages include inline images and attachments
 - [Mailbox] Added new prop 'thread_click_action' that allows to specify the action on clicking an email thread [#369](https://github.com/nylas/components/pull/369)
+- View/save/send draft messages in email thread [#399](https://github.com/nylas/components/pull/399)
 
 ## Bug Fixes